composite_fk_example/requirements.txt
django==4.2
djangorestframework==3.14.0
django-viewflow==2.0.0b6
psycopg==3.1.8